refactoring the syntax crate breaking down syntax.rs into multiple (private) modules.
