+++ razor-fol/src/syntax/theory.rs
@@ -1,0 +1,68 @@
+/*! Defines first-order theories. */
+
+use anyhow::{Error, Result};
+use std::{convert::TryFrom, fmt};
+
+use super::{Formula, Sig};
+
+/// is a first-order theory, containing a set of first-order formulae.
+pub struct Theory {
+    /// is the signature of the theory, containing constants, function symbols, and predicates.
+    pub signature: Sig,
+
+    /// is the list of first-order formulae in this theory.
+    pub formulae: Vec<Formula>,
+}
+
+impl TryFrom<Vec<Formula>> for Theory {
+    type Error = Error;
+
+    fn try_from(formulae: Vec<Formula>) -> Result<Self> {
+        let signature =
+            Sig::try_from(&formulae).map_err(|e| e.context("invalid theory signature"))?;
+
+        Ok(Theory {
+            signature,
+            formulae,
+        })
+    }
+}
+
+impl fmt::Display for Theory {
+    fn fmt(&self, f: &mut fmt::Formatter) -> Result<(), fmt::Error> {
+        let fs: Vec<String> = self.formulae.iter().map(|t| t.to_string()).collect();
+        write!(f, "{}", fs.join("\n"))
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::syntax::{forall1, forall2, forall3};
+    use crate::test_prelude::*;
+
+    #[test]
+    fn test_formula_to_string() {
+        let formulae = vec![
+            forall1(_x(), x().equals(x())),
+            forall2(_x(), _y(), x().equals(y()).implies(y().equals(x()))),
+            forall3(
+                _x(),
+                _y(),
+                _z(),
+                x().equals(y())
+                    .and(y().equals(z()))
+                    .implies(x().equals(z())),
+            ),
+        ];
+
+        let expected_sig = Sig::try_from(&formulae).unwrap();
+        let expected = "∀ x. (x = x)\n\
+            ∀ x, y. ((x = y) → (y = x))\n\
+            ∀ x, y, z. (((x = y) ∧ (y = z)) → (x = z))";
+
+        let theory = Theory::try_from(formulae).unwrap();
+        assert_eq!(expected, &theory.to_string());
+        assert_eq!(&expected_sig, &theory.signature);
+    }
+}
